requirement.txt
pycryptodome==3.20.0
cryptography>=41.0.0
pynacl~=1.5.0
pycrypto==2.6.1
simple-crypt==4.1.7
M2Crypto==0.38.0
python-gnupg==0.5.0
tlslite-ng==0.8.0
rsa==4.9
passlib==1.7.4
petlib==0.0.49
keyring==24.3.0